add Transaction and TransactionType Controllers

+++ back/src/main/java/com/moneda/back/dto/CreateTransactionHistoryDto.java
@@ -2,11 +2,11 @@ package com.moneda.back.dto;
 
 import lombok.Data;
 
-import java.util.Date;
+import java.math.BigDecimal;
 @Data
 public class CreateTransactionHistoryDto {
     private String cvuUser;
     private String transactionType;
-    private Double amount;
+    private BigDecimal amount;
     private String details;
 }
